Extract SD exporter setup to separate file. Add initiating Datadog exporters in the kit server
